requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
Babel==2.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
deepdiff==6.7.1
dictdiffer==0.9.0
docutils==0.20.1
grpcio==1.60.0
idna==3.7
imagesize==1.4.1
Jinja2==3.1.3
MarkupSafe==2.1.5
ordered-set==4.1.0
packaging==24.0
protobuf==4.25.1
pycparser==2.21
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.17.2
pygnmi==0.8.13
PyYAML==6.0.1
requests==2.31.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinx-tabs==3.4.5
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
typing_extensions==4.12.0
urllib3==2.2.1
